third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/items/row-automatic-minimum-for-auto-expected.txt

This is a testharness.js-based test.
[FAIL] Check that min-height is honored when sizing auto rows.
  assert_equals: expected "40px" but got "auto"
[FAIL] Check that min-height is honored when sizing auto rows and spanning grid items.
  assert_equals: expected "20px 30px" but got "auto auto"
[FAIL] Check the interactions between height and min-height and auto tracks.
  assert_equals: expected "60px" but got "auto"
[FAIL] Check that borders and paddings are considering when computing min sizes.
  assert_equals: expected "50px" but got "auto"
Harness: the test ran to completion.
